+++ BVPVWebServer/Services/RoleService.cs
@@ -11,9 +11,9 @@ public class RoleService
         db.Open();
         var sql = searchString switch
         {
-            "" => $"select * from `Roles` where `RoleID` = ' ' order by `RoleLevel`",
-            "*" => $"select * from `Roles` order by `RoleLevel`",
-            _ => $"select * from `Roles` where `RoleID` like '%{searchString}%' order by `RoleLevel`"
+            "" => $"select * from `Admin-Roles` where `RoleID` = ' ' order by `RoleLevel`",
+            "*" => $"select * from `Admin-Roles` order by `RoleLevel`",
+            _ => $"select * from `Admin-Roles` where `RoleID` like '%{searchString}%' order by `RoleLevel`"
         };
 
         var rdr = db.ExecQuery(sql);
@@ -39,7 +39,7 @@ public class RoleService
         var allRoleIds = new List<string>();
         var db = new MariaDb(appInfo);
         db.Open();
-        var rdr = db.ExecQuery($"select `RoleID` from Roles order by `RoleLevel`, `RoleID`");
+        var rdr = db.ExecQuery($"select `RoleID` from `Admin-Roles` order by `RoleLevel`, `RoleID`");
         if (!rdr!.HasRows) return allRoleIds;
         while (rdr.Read())
         {
@@ -55,7 +55,7 @@ public class RoleService
         var gottenRole = new Role();
         var db = new MariaDb(appInfo);
         db.Open();
-        var rdr = db.ExecQuery($"select * from Roles where `RoleID` = '{role}'");
+        var rdr = db.ExecQuery($"select * from `Admin-Roles` where `RoleID` = '{role}'");
         if (!rdr!.HasRows) return gottenRole;
         while (rdr.Read())
         {
@@ -74,7 +74,7 @@ public class RoleService
         using var db = new MariaDb(appInfo);
         db.Open();
         
-        var sql = $"update `Roles` set `RoleLevel` = {role.RoleLevel}, `ReadOnly` = {role.ReadOnly}, `Enabled` = {role.Enabled} where `RoleID` = '{role.RoleId}';";
+        var sql = $"update `Admin-Roles` set `RoleLevel` = {role.RoleLevel}, `ReadOnly` = {role.ReadOnly}, `Enabled` = {role.Enabled} where `RoleID` = '{role.RoleId}';";
         db.ExecNonQuery(sql);
         if (!db.Success) success = false;
         db.Close();
@@ -87,7 +87,7 @@ public class RoleService
         using var db = new MariaDb(appInfo);
         db.Open();
         
-        var sql = $"insert into `Roles` values ('{role.RoleId}', {role.RoleLevel}, {role.ReadOnly}, {role.Enabled});";
+        var sql = $"insert into `Admin-Roles` values ('{role.RoleId}', {role.RoleLevel}, {role.ReadOnly}, {role.Enabled});";
         db.ExecNonQuery(sql);
         if (!db.Success) success = false;
         db.Close();
@@ -100,7 +100,7 @@ public class RoleService
         using var db = new MariaDb(appInfo);
         db.Open();
         
-        var sql = $"delete from `Roles` where `RoleID` = '{roleId}';";
+        var sql = $"delete from `Admin-Roles` where `RoleID` = '{roleId}';";
         db.ExecNonQuery(sql);
         if (!db.Success) success = false;
         db.Close();
@@ -111,7 +111,7 @@ public class RoleService
     {
         var db = new MariaDb(appInfo);
         db.Open();
-        var rdr = db.ExecQuery($"select * from `Roles` where `RoleID` = '{role}'");
+        var rdr = db.ExecQuery($"select * from `Admin-Roles` where `RoleID` = '{role}'");
         return (rdr is {HasRows: true});
     }
 }
@@ -121,6 +121,5 @@ public class Role
     public string? RoleId { get; set; }
     public int RoleLevel { get; set; }
     public bool ReadOnly { get; set; }
-    
     public bool Enabled { get; set; }
 }
